% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addAttributesDfChrSize.R
\name{addAttributesDfChrSize}
\alias{addAttributesDfChrSize}
\title{addAttributesDfChrSize}
\usage{
addAttributesDfChrSize(
  listOfdfChromSize,
  threshold,
  specialOTUNames,
  centromereSize,
  MbThreshold,
  cenFactor,
  chrWidth,
  specialChrWidth,
  squareness,
  modifyChr = TRUE,
  mymessage = TRUE
)
}
\value{
list of data.frames
}
\description{
returns all cols. as character
}
\keyword{internal}
