% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addNotes.R
\name{addNotes}
\alias{addNotes}
\title{This is an internal function that adds notes}
\usage{
addNotes(
  notes,
  listOfdfChromSizenoNA,
  groupName,
  indexCount,
  morphoCount,
  xmnoNA,
  ymnoNA,
  distTextChr,
  chrIdCount,
  notesPosX,
  notesPosY,
  notesTextSize,
  defaultFontFamily2,
  OTUasNote,
  OTUfont,
  OTUfamily,
  downNote,
  rightN,
  leftNoteFont,
  leftNoteFontUp,
  noteFont,
  parseTypes,
  parseStr2lang
)
}
\arguments{
\item{notes}{x component of polygon}

\item{listOfdfChromSizenoNA}{y component of polygon}

\item{groupName}{y of chr}

\item{indexCount}{colors for marks}

\item{morphoCount}{list of df. of mark pos.}

\item{xmnoNA}{thick of border line}

\item{ymnoNA}{boolean}

\item{distTextChr}{numeric}

\item{chrIdCount}{numeric}

\item{notesPosX}{numeric}

\item{notesPosY}{numeric}

\item{notesTextSize}{numeric}

\item{defaultFontFamily2}{character, regex}

\item{downNote}{numeric}
}
\value{
plot
}
\description{
This is an internal function that adds notes
}
\keyword{internal}
