% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrowPlotMark.R
\name{arrowPlotMark}
\alias{arrowPlotMark}
\title{arrowPlotMark
This is an internal function that plot marks}
\usage{
arrowPlotMark(
  squareness,
  xMark,
  yMark,
  dfMarkColorInt,
  listOfdfMarkPosArrow,
  chrWidth,
  n,
  lwd.chr,
  circularPlot,
  y,
  x,
  markLabelSize,
  separFactor,
  labelSpacing,
  circleCenter,
  circleCenterY,
  radius,
  ylistTransChr,
  rotation,
  arrowheadWidthShrink
)
}
\arguments{
\item{squareness}{squareness of vertices <}

\item{xMark}{x component of polygon}

\item{yMark}{yMark component of polygon}

\item{dfMarkColorInt}{colors for marks}

\item{listOfdfMarkPosArrow}{list of df. of mark pos.}

\item{chrWidth}{numeric, width of chr.}

\item{n}{numeric, to define vertices of rounded portions}

\item{lwd.chr}{thick of border line}

\item{circularPlot}{boolean TRUE for circ.}

\item{y}{list, y coords.}

\item{markLabelSize}{numeric, font size}

\item{separFactor}{numeric, sep. kars.}

\item{labelSpacing}{numeric, spacing}

\item{circleCenter}{numeric}

\item{circleCenterY}{numeric}

\item{radius}{numeric}

\item{ylistTransChr}{list, transf. coords.}

\item{rotation}{rotate*}
}
\value{
plot
}
\description{
It returns a plot
}
\keyword{internal}
