% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asymmetryA2.R
\name{asymmetryA2}
\alias{asymmetryA2}
\title{Function to calculate karyotype asymmetry}
\usage{
asymmetryA2(dfChrSize)
}
\arguments{
\item{dfChrSize}{name of dataframe}
}
\value{
list
}
\description{
This function reads a dataframe  with columns:
\code{shortArmSize} and \code{longArmSize} and optionally \code{OTU} and
returns a list with the A2 values for the OTUs

If several species present, use column "OTU".
}
\examples{
asymmetryA2(dfOfChrSize)
asymmetryA2(dfChrSizeHolo)
}
\references{
Romero-Zarco. 1986. A New Method for Estimating Karyotype
Asymmetry. Taxon Vol. 35, No. 3  pp. 526-530
}
\keyword{arm}
\keyword{dataframe}
\keyword{size}
