% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badge_custom.R
\name{badge_custom}
\alias{badge_custom}
\title{Function imported from package badger which is not available in CRAN}
\usage{
badge_custom(x, y, color, url = NULL)
}
\arguments{
\item{x}{x}

\item{y}{y}

\item{color}{color}

\item{url}{url}
}
\value{
character
}
\description{
imported from package badger
}
\keyword{internal}
