% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bigdfChrSizeHolo}
\alias{bigdfChrSizeHolo}
\title{Example data for holocentrics for several species}
\format{A data frame:
\describe{
\item{OTU}{OTU, species}
\item{chrName}{name of chromosome}
\item{chrSize}{size of chromosome, micrometers or Mb}
}}
\usage{
bigdfChrSizeHolo
}
\description{
A dataset containing chromosome basic data
}
\keyword{datasets}
