% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bigdfDataCen}
\alias{bigdfDataCen}
\title{Example data for centromeric marks df with column OTU}
\format{A data frame with 53940 rows and 10 variables:
\describe{
\item{OTU}{OTU, species}
\item{chrName}{name of chromosome}
\item{markName}{name of mark (site)}
}}
\usage{
bigdfDataCen
}
\description{
A dataset containing the prices and other attributes of almost 54,000
diamonds.
}
\keyword{datasets}
