% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bigdfMarkPosHolo}
\alias{bigdfMarkPosHolo}
\title{Example data for mark position of holocentrics with column OTU}
\format{A data frame:
\describe{
\item{OTU}{OTU, species}
\item{chrName}{name of chromosome}
\item{markName}{name of mark}
\item{markPos}{position from bottom or top (see parameter origin)}
\item{markSize}{size of mark in micrometers or Mb}
}}
\usage{
bigdfMarkPosHolo
}
\description{
A dataset containing mark position
}
\keyword{datasets}
