% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bigdfOfChrSize}
\alias{bigdfOfChrSize}
\title{Example data for monocentrics for several species}
\format{A data frame:
\describe{
\item{OTU}{OTU, species}
\item{chrName}{name of chromosome}
\item{shortArmSize}{size of short arm, micrometers }
\item{longArmSize}{size of long arm, micrometers }
}}
\usage{
bigdfOfChrSize
}
\description{
A dataset containing chromosome basic data
}
\keyword{datasets}
