% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bigdfOfMarks}
\alias{bigdfOfMarks}
\title{Example data for mark position of monocentrics with column OTU}
\format{A data frame:
\describe{
\item{OTU}{OTU, species}
\item{chrName}{name of chromosome}
\item{markName}{name of mark}
\item{markArm}{use p for short and q for long}
\item{markDistCen}{distance of mark to centromere}
\item{markSize}{size of mark in micrometers or Mb}
}}
\usage{
bigdfOfMarks
}
\description{
A dataset containing mark data
}
\keyword{datasets}
