% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cenmarkdata.R
\docType{data}
\name{cenmarkdata}
\alias{cenmarkdata}
\alias{bigdfDataCen}
\alias{dfOfCenMarks}
\alias{allCenMarksSample}
\title{Centromeric Mark data}
\format{bigdfDataCen a data.frame with cols:
\describe{
\item{OTU}{OTU, species, mandatory if in dfChrSize}
\item{chrName}{name of chromosome}
\item{markName}{name of mark (site)}
}}
\usage{
bigdfDataCen

dfOfCenMarks

allCenMarksSample
}
\description{
bigdfDataCen: Example data for centromeric marks df with column OTU

dfOfCenMarks: Example data for centromeric marks

allCenMarksSample: Example data for centromeric marks
}
\seealso{
\code{\link{plotIdiograms}}

\code{\link{dfMarkColor}}

\code{\link{markpos}}

\code{\link{armRatioCI}}
}
\keyword{datasets}
