% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkArmHolocenError.R
\name{checkArmHolocenError}
\alias{checkArmHolocenError}
\title{checkArmHolocenError
This is an internal function that makes list of d.fs}
\usage{
checkArmHolocenError(listOfdfArmGISHInternal, CenTypeNames)
}
\arguments{
\item{listOfdfArmGISHInternal}{df list}

\item{CenTypeNames}{character vector}
}
\value{
list
}
\description{
checkArmHolocenError
This is an internal function that makes list of d.fs
}
\keyword{internal}
