% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkNameOTUdfMarks.R
\name{checkNameOTUdfMarks}
\alias{checkNameOTUdfMarks}
\title{Function to check names of chromosomes among data.frames}
\usage{
checkNameOTUdfMarks(dfChrSize, dfMarks)
}
\arguments{
\item{dfChrSize}{data.frame, of chr. size with columns: OTU (optional),
chrName, for chr. with cen.: shortArmSize, longArmSize, for holoc.: chrSize}

\item{dfMarks}{data.frame, of marks (sites): OTU (opt / mandat. if in
dfChrSize), chrName markName markSize, for chr. with cen.:  chrRegion (p,q)
markDistCen, for holoc: markPos}
}
\value{
logical
}
\description{
This function reads two data.frames the one with chromosome sizes
\code{dfChrSize} and one df with data of marks (sites) both have to have the
column \code{chrName} and if several species, \code{OTU}. The functions
returns \code{TRUE} if names of OTUs in data.frame of marks are included in
data.frame of chr. size.
}
\keyword{check}
\keyword{data.frame}
\keyword{internal}
