% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrbasicdataHolo.R
\docType{data}
\name{chrbasicdataHolo}
\alias{chrbasicdataHolo}
\alias{dfChrSizeHolo}
\alias{bigdfChrSizeHolo}
\title{Chr. basic data Holo.}
\format{data.frame with cols:
\describe{
\item{OTU}{grouping OTU (species), optional if only one OTU}
\item{chrName}{name of chromosome}
\item{chrSize}{size of chromosome, micrometers or Mb}
\item{group}{chromosome group, optional}
}}
\usage{
dfChrSizeHolo

bigdfChrSizeHolo
}
\description{
dfChrSizeHolo: Example data for holocentrics for 1 species

bigdfChrSizeHolo: Example data for holocentrics for several species, OTU
}
\seealso{
\code{\link{asymmetryA2}}

\code{\link{plotIdiograms}}

\code{\link{markdataholo}}
}
\keyword{datasets}
