% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cMPlotMark.R
\name{cMPlotMark}
\alias{cMPlotMark}
\title{This is an internal function that eliminates factors}
\usage{
cMPlotMark(
  xMark,
  yMark,
  y,
  x,
  dfMarkColorInternal,
  listOfdfMarkPoscM,
  lwd.cM,
  circularPlot,
  radius,
  circleCenter,
  circleCenterY,
  separFactor,
  markLabelSize,
  pattern,
  n,
  labelSpacing,
  chrWidth,
  ylistTransChr,
  rotation,
  labelOutwards
)
}
\arguments{
\item{xMark}{x component of polygon}

\item{yMark}{y component of polygon}

\item{y}{y of chr}

\item{dfMarkColorInternal}{colors for marks}

\item{listOfdfMarkPoscM}{list of df. of mark pos.}

\item{lwd.cM}{thick of border line}

\item{circularPlot}{boolean}

\item{radius}{numeric}

\item{circleCenter}{numeric}

\item{circleCenterY}{numeric}

\item{separFactor}{numeric}

\item{markLabelSize}{numeric}

\item{pattern}{character, regex}

\item{n}{numeric}

\item{labelSpacing}{numeric}

\item{chrWidth}{numeric, chr. width}

\item{ylistTransChr}{list}

\item{rotation}{rotation}

\item{labelOutwards}{srt}
}
\value{
plot
}
\description{
It returns a data.frames
}
\keyword{internal}
