% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customCeiling.R
\name{customCeiling}
\alias{customCeiling}
\title{Function to do ceiling at different levels}
\usage{
customCeiling(x, customDecimals = 1)
}
\arguments{
\item{x}{number to round}

\item{level}{level of ceiling}
}
\value{
numeric
}
\description{
This function reads number to round
}
\keyword{internal}
