% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfChrSizeHolo}
\alias{dfChrSizeHolo}
\title{Example data for holocentrics for 1 species}
\format{A data frame:
\describe{
\item{chrName}{name of chromosome}
\item{chrSize}{size of chromosome, micrometers or Mb}
}}
\usage{
dfChrSizeHolo
}
\description{
A dataset containing chromosome basic data
}
\keyword{datasets}
