% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfMarkColor}
\alias{dfMarkColor}
\title{Example General data for marks}
\format{A data frame: \describe{ \item{markName}{name of mark}
\item{markColor}{use R colors} \item{style}{character, use square or dots}
}}
\usage{
dfMarkColor
}
\description{
A dataset containing mark general data, not position
}
\keyword{datasets}
