% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfMarkPosHolo}
\alias{dfMarkPosHolo}
\title{Example data for mark position of holocentrics}
\format{A data frame:
\describe{
\item{chrName}{name of chromosome}
\item{markName}{name of mark}
\item{markPos}{position from bottom or top (see parameter origin)}
\item{markSize}{size of mark in micrometers or Mb}
}}
\usage{
dfMarkPosHolo
}
\description{
A dataset containing mark position
}
\keyword{datasets}
