% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfOfCenMarks}
\alias{dfOfCenMarks}
\title{Example data for centromeric marks of monocentrics}
\format{A data frame:
\describe{
\item{chrName}{name of chromosome}
\item{markName}{name of mark}
}}
\usage{
dfOfCenMarks
}
\description{
A dataset containing centromeric mark
}
\keyword{datasets}
