% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dfOfChrSize}
\alias{dfOfChrSize}
\title{Example data for monocentrics for several species}
\format{A data frame:
\describe{
\item{chrName}{name of chromosome}
\item{shortArmSize}{size of short arm, micrometers }
\item{longArmSize}{size of long arm, micrometers }
}}
\usage{
dfOfChrSize
}
\description{
A dataset containing chromosome basic data
}
\keyword{datasets}
