% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfToListColumn.R
\name{dfToListColumn}
\alias{dfToListColumn}
\title{dfToListColumn
This is an internal function that makes list of d.fs}
\usage{
dfToListColumn(dfLookColumn, columnToSplit = "OTU")
}
\arguments{
\item{dfLookColumn}{df with or without Column to list}

\item{columnToSplit}{character, column name to split}
}
\value{
list
}
\description{
dfToListColumn
This is an internal function that makes list of d.fs
}
\keyword{internal}
