% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfToListOTU.R
\name{dfToListOTU}
\alias{dfToListOTU}
\title{dfToListOTU
This is an internal function that makes list of d.fs}
\usage{
dfToListOTU(dfLookOTU)
}
\arguments{
\item{dfLookOTU}{df with or without otu to list}
}
\value{
list
}
\description{
dfToListOTU
This is an internal function that makes list of d.fs
}
\keyword{internal}
