% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterExtraOTU.R
\name{filterExtraOTU}
\alias{filterExtraOTU}
\title{filterExtraOTU
This is an internal function that filters list of d.f. of marks}
\usage{
filterExtraOTU(listOfdfChromSize, listOfdfMarkPosCenType)
}
\arguments{
\item{listOfdfChromSize}{list of all d.f. of chr.}

\item{listOfdfMarkPosCenType}{only d.f. of this cen. type}
}
\value{
data.frame
}
\description{
It returns rulers (axis)
}
\keyword{internal}
