% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixChrNameDup.R
\name{fixChrNameDup}
\alias{fixChrNameDup}
\title{This is an internal function fixes duplicated names when no marks}
\usage{
fixChrNameDup(listOfdfChromSize, mybooleanChrName)
}
\arguments{
\item{listOfdfChromSize}{list of df of chr size}

\item{mybooleanChrName}{test for absence of marks}
}
\value{
Returns a graphics element
}
\description{
It returns a list of dataframes
}
\keyword{internal}
