% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixChrNameDupDF.R
\name{fixChrNameDupDF}
\alias{fixChrNameDupDF}
\title{This is an internal function fixes duplicated names when no marks}
\usage{
fixChrNameDupDF(list1dfChromSize, mybooleanChrName)
}
\arguments{
\item{list1dfChromSize}{list of 1 df of chr size}

\item{mybooleanChrName}{test for absence of marks}
}
\value{
Returns a graphics element
}
\description{
It returns a list of data.frames
}
\keyword{internal}
