% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genBankReadIF.R
\name{genBankReadIF}
\alias{genBankReadIF}
\title{FUNCTION genBankReadIF}
\usage{
genBankReadIF(filename.gb, forbiddenFields = "translation")
}
\arguments{
\item{filename.gb}{name of file to read, downloaded from genBank, or,
object from \code{rentrez::entrez_fetch( db="nuccore",} \code{ id="theID",} \code{ rettype="gbwithparts",} \code{ retmode = "text" )}}

\item{forbiddenFields}{names of field of feature (CDS, gene) to ignore, separated by |. Defaults to \code{"translation"}}
}
\value{
list
}
\description{
genBankReadIF: creates a list of data.frames from a genbank
format file or a
rentrez object

Requires tidyr
}
\keyword{genBank}
