% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertInList.R
\name{insertInList}
\alias{insertInList}
\title{This is an internal function that inserts in vector or list}
\usage{
insertInList(list, pos, elems)
}
\arguments{
\item{list}{vector or list}

\item{pos}{index to insert in}

\item{elems}{elements to be inserted}
}
\value{
list or vector
}
\description{
It returns list or vector
}
\keyword{internal}
