% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.uniqueIF.R
\name{make.uniqueIF}
\alias{make.uniqueIF}
\title{FUNCTION make.uniqueIF}
\usage{
make.uniqueIF(string, sep = "_", letter = FALSE)
}
\arguments{
\item{string}{name of char. vector}

\item{sep}{separator}

\item{letter}{boolean, use numbers when \code{FALSE}}
}
\value{
character vector
}
\description{
make.uniqueIF: produces unique strings from duplicated
}
\examples{
make.uniqueIF(c(1, 1, 2, 2))

}
\keyword{duplicates}
\keyword{string}
\keyword{unique}
