% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makedfMarkColorMycolors.R
\name{makedfMarkColorMycolors}
\alias{makedfMarkColorMycolors}
\title{FUNCTION to make a data.frame of Marks' characterisitcs}
\usage{
makedfMarkColorMycolors(
  markNames,
  mycolors,
  colorstoremove = NULL,
  defaultStyleMark = "square"
)
}
\arguments{
\item{markNames}{names of marks}

\item{mycolors}{character vector of names of colors}

\item{colorstoremove}{character vector of colors to remove from mycolors. Default \code{NULL}}

\item{defaultStyleMark}{character vector with default style \code{"square"}. Other options:  \code{"squareLeft"}, \code{"cM"}
\code{"cMLeft"}, \code{"dots"}, \code{"upArrow"}, \code{"downArrow"}, \code{"cenStyle"}}
}
\value{
data.frame
}
\description{
This function reads character vector with marks names, a
character vector of
prohibited colors and a custom list of colors
to be associated with the mark names

It returns a data.frame with color and style for each mark
}
\keyword{create}
\keyword{mark}
