% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNumCols.R
\name{makeNumCols}
\alias{makeNumCols}
\title{These are internal functions that eliminate factors}
\usage{
makeNumCols(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame
}
\description{
It returns a data.frame
}
\keyword{internal}
