% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeVectorNames.R
\name{makeVectorNames}
\alias{makeVectorNames}
\title{makeVectorNames
This is an internal function that makes a vector of names}
\usage{
makeVectorNames(namedList, attr, selector)
}
\arguments{
\item{namedList}{original list}

\item{attr}{attribute}

\item{selector}{value of attribute}
}
\value{
character vector
}
\description{
makeVectorNames
This is an internal function that makes a vector of names
}
\keyword{internal}
