% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGGChr.R
\name{mapGGChr}
\alias{mapGGChr}
\alias{mapGGChrMark}
\title{FUNCTIONS mapGGChr and mapGGChrMark (for ggplot)}
\usage{
mapGGChr(dfChrSize, chrSpacing = 0.5)

mapGGChrMark(dfChrSize, dfMarkPos, chrSpacing = 0.5)
}
\arguments{
\item{dfChrSize}{character, name of data.frame with columns: chrSize}

\item{chrSpacing}{numeric, \code{1 - chrSpacing} will be the width of chr.}

\item{dfMarkPos}{(\code{mapGGChrMark}) name of data.frame of marks}
}
\value{
list

list
}
\description{
Currently works for holocentrics when only one OTU. See vignette.

mapGGChr: reads a data.frame and produces coordinates for ggplot of chr.

mapGGChrMark: reads data.frames and produces data.frames for ggplot of chr. and marks
}
\keyword{arm}
\keyword{data.frame}
\keyword{size}
