% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapXY.R
\name{mapXY}
\alias{mapXY}
\title{mapXY
This is an internal function that creates coords for chr. or arms or marks}
\usage{
mapXY(start, end, y, yMod, x, yfactor, r2, pts_1, pts_2, pts_3, pts_4)
}
\arguments{
\item{start}{1st index}

\item{end}{last index}

\item{y}{y coords}

\item{yMod}{y coords modified because of squareness}

\item{x}{x coords}

\item{yfactor}{y x factor}

\item{r2}{radius}

\item{pts_1}{points of squareness}

\item{pts_2}{points of squareness}

\item{pts_3}{points of squareness}

\item{pts_4}{points of squareness}

\item{n}{number of edges in rounded vertices}
}
\value{
list
}
\description{
mapXY
This is an internal function that creates coords for chr. or arms or marks
}
\keyword{internal}
