% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapXYCen.R
\name{mapXYCen}
\alias{mapXYCen}
\title{mapXYCen
This is an internal function that creates coords for chr. or arms or marks}
\usage{
mapXYCen(start, end, y, yMod, x, yfactor, r2, pts)
}
\arguments{
\item{start}{1st index}

\item{end}{last index}

\item{y}{y coords}

\item{yMod}{y coords modified because of roundness}

\item{yfactor}{y x factor}

\item{r2}{radius}

\item{pts}{points of roundness}
}
\value{
list
}
\description{
mapXYCen
This is an internal function that creates coords for chr. or arms or marks
}
\keyword{internal}
