% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdataholo.R
\docType{data}
\name{markdataholo}
\alias{markdataholo}
\alias{bigdfMarkPosHolo}
\alias{dfMarkPosHolo}
\alias{dfAlloParentMarksHolo}
\alias{bigdfOfMarks3Mb}
\alias{bigdfMarkPosHolo2}
\title{Mark Positional data - Holocen.}
\format{
data.frame with columns:
\describe{
\item{OTU}{OTU, species, optional}
\item{chrName}{name of chromosome}
\item{markName}{name of mark}
\item{markPos}{position from bottom or top (see parameter \code{origin}
in plotIdiograms)}
\item{markSize}{size of mark in micrometers or Mb}
}

An object of class \code{data.frame} with 6 rows and 4 columns.

An object of class \code{data.frame} with 13 rows and 4 columns.

An object of class \code{data.frame} with 12 rows and 5 columns.

An object of class \code{data.frame} with 18 rows and 6 columns.
}
\usage{
bigdfMarkPosHolo

dfMarkPosHolo

dfAlloParentMarksHolo

bigdfOfMarks3Mb

bigdfMarkPosHolo2
}
\description{
When several OTUs, some can be monocen. and others holocen.
Marks distance for
monocen. are measured from cen. and for
holocen. from top or bottom depending on \code{param} \code{origin}. See
vignettes.

bigdfMarkPosHolo: Example data for mark position of
holocentrics with column OTU

dfMarkPosHolo: Example data for mark position of holocentrics

dfAlloParentMarksHolo: Example data for mark position of GISH

bigdfOfMarks3Mb: Example data for mark position in Mb

bigdfMarkPosHolo2: Example data for mark position
}
\seealso{
\code{\link{markposDFs}}

\code{\link{plotIdiograms}}

\code{\link{chrbasicdataHolo}}
}
\keyword{datasets}
