% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markDistCenGISHfix.R
\name{markDistCenGISHfix}
\alias{markDistCenGISHfix}
\title{markDistCenGISHfix
This is an internal function that fixes markDistCen when GISH}
\usage{
markDistCenGISHfix(
  dfArmGISHInternalMonocen,
  dfChrSizeInternal,
  columnArmSize,
  markDistType
)
}
\arguments{
\item{dfArmGISHInternalMonocen}{df of GISH OF ARM}

\item{dfChrSizeInternal}{d.f. of chr. size}

\item{columnArmSize}{column rel. to arm.}

\item{markDistType}{character}
}
\value{
data.frame
}
\description{
markDistCenGISHfix
This is an internal function that fixes markDistCen when GISH
}
\keyword{internal}
