% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markpos.R
\docType{data}
\name{markpos}
\alias{markpos}
\alias{bigdfOfMarks}
\alias{dfOfMarks}
\alias{humMarkPos}
\title{Mark Positional data - monocentrics}
\format{bigdfOfMarks a data.frame with cols:
\describe{
\item{OTU}{OTU, species, mandatory if in dfChrSize}
\item{chrName}{name of chromosome}
\item{markName}{name of mark}
\item{markArm}{use p for short and q for long}
\item{markDistCen}{distance of mark to centromere}
\item{markSize}{size of mark}
}}
\source{
\href{http://www.pathology.washington.edu/research/cytopages/idiograms/human/}{Washington U}
}
\usage{
bigdfOfMarks

dfOfMarks

humMarkPos
}
\description{
bigdfOfMarks: Example data for mark position with column OTU

dfOfMarks: Example data for mark position

humMarkPos: human karyotype mark position
}
\references{
Adler 1994. Idiogram Album. URL: \href{http://www.pathology.washington.edu/research/cytopages/idiograms/human/}{Washington U.}
}
\seealso{
\code{\link{plotIdiograms}}

\code{\link{chrbasicdatamono}}

\code{\link{dfMarkColor}}
}
\keyword{datasets}
