% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newOrderColumn.R
\name{newOrderColumn}
\alias{newOrderColumn}
\title{newOrderColumn
This is an internal function that adds neworder column to listOfdfChromSizeCenType}
\usage{
newOrderColumn(listOfdfChromSize, listOfdfChromSizeCenType)
}
\arguments{
\item{listOfdfChromSize}{list of all d.f. of chr.}

\item{listOfdfChromSizeCenType}{only d.f. of this cen. type}
}
\value{
data.frame
}
\description{
It returns rulers (axis)
}
\keyword{internal}
