% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perMark.R
\name{perMark}
\alias{perMark}
\title{FUNCTION perMark}
\usage{
perMark(dfMarkPos, listOfdfChromSize, result = "list", bToRemove = "")
}
\arguments{
\item{dfMarkPos}{data.frame, of marks' position}

\item{listOfdfChromSize}{list of data.frames or data.frame of chr. sizes.}

\item{result}{character, type of return, \code{"data.frame"} or \code{"list"}}

\item{bToRemove}{character vector, bands to ignore}
}
\value{
list
}
\description{
calculates fraction (\%) of chromosome for each mark
}
\examples{
load(system.file("shinyApps", "iBoard/www/rda/monoholoCS.rda", package = "idiogramFISH"))
load(system.file("shinyApps", "iBoard/www/rda/monoholoMarks.rda", package = "idiogramFISH"))
monoholoMarks2 <- fillMarkInfo(monoholoMarks, monoholoCS)
perMark(monoholoMarks2, monoholoCS, result = "data.frame")

}
\keyword{chromosome}
\keyword{mark}
\keyword{percentage}
\keyword{span}
