% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDotMarks.R
\name{plotDotMarks}
\alias{plotDotMarks}
\title{plotDotMarks
This is an internal function that plot dot marks}
\usage{
plotDotMarks(
  bannedMarkName,
  xMarkCr,
  yMarkCr,
  rad,
  colCr,
  n,
  xfactor,
  colBorderCr,
  circularPlot,
  y,
  x,
  radius,
  circleCenter,
  circleCenterY,
  separFactor,
  chrWidth,
  listOfdfMarkPosCr,
  markLabelSize,
  pattern,
  labelSpacing,
  useOneDot,
  legend,
  ylistTransChr,
  rotation,
  labelOutwards
)
}
\arguments{
\item{xMarkCr}{x coords}

\item{yMarkCr}{y coords}

\item{rad}{radius}

\item{colCr}{color}

\item{n}{vertices number}

\item{xfactor}{aspect}

\item{colBorderCr}{border color}

\item{circularPlot}{boolean}

\item{y}{y of chr.}

\item{x}{x of chr.}

\item{radius}{numeric}

\item{circleCenter}{numeric}

\item{circleCenterY}{numeric}

\item{separFactor}{numeric}

\item{chrWidth}{numeric}

\item{listOfdfMarkPosCr}{list of d.f.}

\item{markLabelSize}{numeric}

\item{pattern}{numeric}

\item{labelSpacing}{numeric}

\item{useOneDot}{boolean}

\item{legend}{character}

\item{ylistTransChr}{list of trans. coords}

\item{rotation}{rotation}

\item{labelOutwards}{srt}
}
\value{
plot
}
\description{
It returns a plot
}
\keyword{internal}
