% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotlabelsright.R
\name{plotlabelsright}
\alias{plotlabelsright}
\title{This is an internal function to plot labels to the right, when "aside" for its position}
\usage{
plotlabelsright(x, y, markLabelSpacer, chrWidth, dfMarkColorInternal,
  allMarkMaxSize, normalizeToOne, markLabelSize, xfactor)
}
\arguments{
\item{x}{The x axis coordinates of chromosomes}

\item{y}{The y axis coordinates of chromosomes}

\item{markLabelSpacer}{distance from right chr to legend}

\item{chrWidth}{chr widht}

\item{dfMarkColorInternal}{dataframe of mark characteristics}

\item{allMarkMaxSize}{maximum size of marks}

\item{normalizeToOne}{transformation value of karyotype height}

\item{markLabelSize}{font size of legends}

\item{xfactor}{relative proportion of x vs y axes}
}
\value{
Returns a graphics element
}
\description{
It returns a graphic element with the legends
}
\keyword{internal}
