% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sqPlotMarkCP.R
\name{plotSqMarks}
\alias{plotSqMarks}
\title{sqPlotMarkCP
This is an internal function that plot marks}
\usage{
plotSqMarks(
  chromatidsCheck_1,
  chromatidsCheck_2,
  circularPlot,
  squareness,
  xMarkSq,
  yMarkSq,
  dfMarkColorInt,
  lwd.marks2,
  listMarkPosSq,
  listChrSize,
  specialChrWidth,
  chrWidth,
  yfactor,
  markN,
  ptsl,
  bannedMarkName3,
  y,
  markLabelSize,
  pattern,
  separFactor,
  labelSpacing,
  circleCenter,
  circleCenterY,
  radius,
  legend,
  ylistTransChrSimple,
  rotation,
  labelOutwards,
  x,
  xModifierMH,
  pts,
  alpha_val = 1
)
}
\arguments{
\item{circularPlot}{boolean TRUE for circ.}

\item{squareness}{squareness of vertices <}

\item{dfMarkColorInt}{colors for marks}

\item{listMarkPosSq}{list of df. of mark pos.}

\item{listChrSize}{chr size df list}

\item{specialChrWidth}{numeric, width of chr.}

\item{chrWidth}{numeric, width of chr.}

\item{yfactor}{yMark distortion based on canvas proportion}

\item{y}{list, y coords.}

\item{markLabelSize}{numeric, font size}

\item{pattern}{character, regex}

\item{separFactor}{numeric, sep. kars.}

\item{labelSpacing}{numeric, spacing}

\item{circleCenter}{numeric}

\item{circleCenterY}{numeric}

\item{radius}{numeric}

\item{legend}{character}

\item{rotation}{rotate}

\item{labelOutwards}{srt}

\item{xMark}{x component of polygon}

\item{yMark}{yMark component of polygon}

\item{n}{numeric, to define vertices of rounded portions}

\item{lwd.chr}{thick of border line}

\item{ylistTransChr}{list, transf. coords.}
}
\value{
plot
}
\description{
It returns a plot
}
\keyword{internal}
