% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeNAFromList.R
\name{removeNAFromList}
\alias{removeNAFromList}
\title{removeNAFromList
This is an internal function that removes NA from list}
\usage{
removeNAFromList(noNA, areNA)
}
\arguments{
\item{noNA}{original list}

\item{areNA}{which are NA}

\item{mycolors}{vector of colors}
}
\value{
list
}
\description{
It returns rulers (axis)
}
\keyword{internal}
