% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundPlot.R
\name{roundPlot}
\alias{roundPlot}
\title{This is an internal function that eliminates factors}
\usage{
roundPlot(roundness, x, y, col, border, chrWidth, yfactor, n2, lwd.chr)
}
\arguments{
\item{roundness}{round}

\item{x}{x}

\item{y}{y}

\item{col}{col}

\item{border}{border}

\item{chrWidth}{width of}

\item{yfactor}{y distortion}

\item{n2}{data.frame}

\item{lwd.chr}{thick of border}
}
\value{
plot
}
\description{
It returns a dataframes
}
\keyword{internal}
