% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulerTitle.R
\name{rulerTitle}
\alias{rulerTitle}
\title{rulerTitle
This is an internal function that plots rulers}
\usage{
rulerTitle(
  xmnoNA,
  ymnoNA,
  chrSpacing,
  yPosRulerTitle,
  listOfdfChromSizeCenType,
  MbUnit,
  specialyTitle,
  yTitle,
  xPosRulerTitle,
  rulerTitleSize
)
}
\arguments{
\item{xmnoNA}{x coords}

\item{ymnoNA}{y coords}

\item{chrSpacing}{chr. spacing}

\item{listOfdfChromSizeCenType}{only d.f. of this cen. type}

\item{MbUnit}{intervals}

\item{specialyTitle}{modifier of pos.}

\item{yTitle}{pos. of ruler}

\item{xPosRulerTitle}{font}

\item{rulerTitleSize}{font size}
}
\value{
plot
}
\description{
It returns rulers (axis)
}
\keyword{internal}
