% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBoard.R
\name{runBoard}
\alias{runBoard}
\title{FUNCTION runBoard}
\usage{
runBoard(installAll = FALSE)
}
\arguments{
\item{installAll}{boolean, when \code{TRUE} dependences are installed without asking. Defaults to \code{FALSE}}
}
\value{
shiny
}
\description{
runBoard: run shinyApp
}
\keyword{shiny}
