% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textLabelDots.R
\name{textLabelDots}
\alias{textLabelDots}
\title{textLabelDots
This is an internal function that plot labels of cM marks}
\usage{
textLabelDots(
  xMarkCr,
  yMarkCr,
  listOfdfChromSize,
  listOfdfMarkPosCr,
  specialChrSpacing,
  chrSpacing,
  markLabelSize,
  pattern,
  xBoundariesQuar
)
}
\arguments{
\item{listOfdfChromSize}{radius}

\item{listOfdfMarkPosCr}{color}

\item{specialChrSpacing}{vertices number}

\item{chrSpacing}{aspect}

\item{markLabelSize}{size of mark label font}

\item{pattern}{pattern to remove from mark name}

\item{xBoundariesQuar}{x size}

\item{xMark}{x coords}

\item{yMark}{y coords}
}
\value{
plotted text
}
\description{
It returns a plot
}
\keyword{internal}
