% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientation.R
\name{yVertoHor}
\alias{yVertoHor}
\title{FUNCTIONS: yVertoHor, xHortoVer, xHortoVerRoundCen, xHortoVerDots,
mapCircle, mapRadius,
transYList, mapOTUnames, addOTUnames, applyMapCircle, intercalate,
drawPlot, drawPlotMark,
drawPlotMarkLine, drawCen, circLabelMark, circPlotDots, plotChrNames,
addChrNameAttrMark,
addChrNameAttrMarkDots, xMarkMap, xMarkMapLeft, markMapPer, markMapPercM,
markMapPerDots,
radDotsPer, centerMarkMapPer, transyListMark, transyListCen, transRadDots,
transyListMarkDots,
oneDot, OTUlabelsright}
\usage{
yVertoHor(y, monocenNames)
}
\arguments{
\item{y}{list, chr coords.}

\item{monocenNames}{character, vector with names of OTUs}
}
\value{
plot
}
\description{
FUNCTIONS: yVertoHor, xHortoVer, xHortoVerRoundCen, xHortoVerDots,
mapCircle, mapRadius,
transYList, mapOTUnames, addOTUnames, applyMapCircle, intercalate,
drawPlot, drawPlotMark,
drawPlotMarkLine, drawCen, circLabelMark, circPlotDots, plotChrNames,
addChrNameAttrMark,
addChrNameAttrMarkDots, xMarkMap, xMarkMapLeft, markMapPer, markMapPercM,
markMapPerDots,
radDotsPer, centerMarkMapPer, transyListMark, transyListCen, transRadDots,
transyListMarkDots,
oneDot, OTUlabelsright
}
\keyword{internal}
