% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armRatioCI.R
\name{armRatioCI}
\alias{armRatioCI}
\title{Function to calculate Arm Ratio and C.I. and morphological categories}
\usage{
armRatioCI(dfOfChrSize)
}
\arguments{
\item{dfOfChrSize}{name of dataframe with columns: shortArmSize, longArmSize}
}
\value{
data.frame
}
\description{
This function reads a dataframe and produces AR (r), CI, Guerra and Levan
classifications
}
\examples{
data(dfOfChrSize)
armRatioCI(dfOfChrSize)
}
\references{
LEVAN A, FREDGA K, SANDBERG AA (1964) NOMENCLATURE FOR
CENTROMERIC POSITION ON CHROMOSOMES. Hereditas 52:201–220.

Guerra. 1986. Reviewing the chromosome nomenclature of Levan et
al. Braz. Jour. Gen. Vol IX, 4, 741-743
}
\keyword{arm}
\keyword{dataframe}
\keyword{size}
