% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrbasicdataHolo.R
\docType{data}
\name{chrbasicdataHolo}
\alias{chrbasicdataHolo}
\alias{dfChrSizeHolo}
\alias{bigdfChrSizeHolo}
\alias{parentalAndHybHoloChrSize}
\alias{bigdfOfChrSize3Mb}
\title{Chr. basic data Holo.}
\format{
data.frame with columns:
\describe{
\item{OTU}{grouping OTU (species), optional if only one OTU}
\item{chrName}{name of chromosome}
\item{chrSize}{size of chromosome, micrometers or Mb}
\item{group}{chromosome group, optional}
}

An object of class \code{data.frame} with 12 rows and 3 columns.

An object of class \code{data.frame} with 13 rows and 3 columns.

An object of class \code{data.frame} with 14 rows and 3 columns.
}
\usage{
dfChrSizeHolo

bigdfChrSizeHolo

parentalAndHybHoloChrSize

bigdfOfChrSize3Mb
}
\description{
dfChrSizeHolo: Example data for holocentrics for 1 species

bigdfChrSizeHolo: Example data for holocentrics for several
species, OTU

parentalAndHybHoloChrSize: Example data for holocentrics for
several species, OTU

bigdfOfChrSize3Mb: Example data in Mb without chr. arms for
three species, OTU
}
\seealso{
\code{\link{asymmetryA2}}

\code{\link{plotIdiograms}}

\code{\link{markdataholo}}
}
\keyword{datasets}
