% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chrbasicdatamono.R
\docType{data}
\name{chrbasicdatamono}
\alias{chrbasicdatamono}
\alias{dfOfChrSize}
\alias{bigdfOfChrSize}
\alias{humChr}
\alias{allChrSizeSample}
\title{Chr. basic data}
\format{data.frame with cols:
\describe{
\item{OTU}{OTU, species, optional if only one OTU (species)}
\item{chrName}{name of chromosome}
\item{shortArmSize}{size of short arm, micrometers }
\item{longArmSize}{size of long arm, micrometers }
\item{group}{chr group, optional}
}}
\source{
\url{http://www.pathology.washington.edu/research/cytopages/idiograms/human/}
}
\usage{
dfOfChrSize

bigdfOfChrSize

humChr

allChrSizeSample
}
\description{
dfOfChrSize: Example data for monocentrics

bigdfOfChrSize: Example data for monocentrics for several species, OTU

humChr: Example data for human karyotype (Adler 1994)

allChrSizeSample: Example data for monocentrics for several species, OTU
}
\references{
Adler 1994. Idiogram Album. \url{http://www.pathology.washington.edu/research/cytopages/idiograms/human/}
}
\seealso{
\code{\link{plotIdiograms}}

\code{\link{armRatioCI}}

\code{\link{asymmetry}}

\code{\link{markpos}}

\code{\link{cenmarkdata}}
}
\keyword{datasets}
