% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfMarkStyle.R
\docType{data}
\name{dfMarkStyle}
\alias{dfMarkStyle}
\alias{dfMarkColor}
\alias{humMarkColor}
\alias{mydfMaColor}
\title{Mark characteristics}
\format{dfMarkColor a data.frame with cols:
\describe{
\item{markName}{name of mark}
\item{markColor}{use R colors}
\item{style}{character, use square or dots, optional}
}}
\source{
\url{http://www.pathology.washington.edu/research/cytopages/idiograms/human/}
}
\usage{
dfMarkColor

humMarkColor

mydfMaColor
}
\description{
dfMarkColor: Example General data for marks NOT position

humMarkColor: human mark characteristics

mydfMaColor: mark characteristics used in vignette of phylogeny
}
\references{
Adler 1994. Idiogram Album. URL: \url{http://www.pathology.washington.edu/research/cytopages/idiograms/human/}
}
\seealso{
\code{\link{plotIdiograms}}

\code{\link{plotIdiogramsHolo}}

\code{\link{markpos}}

\code{\link{markdataholo}}

\code{\link{cenmarkdata}}
}
\keyword{datasets}
