% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfMarkStyle.R
\docType{data}
\name{dfMarkStyle}
\alias{dfMarkStyle}
\alias{dfMarkColor}
\alias{humMarkColor}
\alias{mydfMaColor}
\alias{dfMarkColor5S25S}
\title{Mark characteristics}
\format{
dfMarkColor a data.frame with columns:
\describe{
\item{markName}{name of mark}
\item{markColor}{use R colors}
\item{style}{character, use square or dots, optional}
\item{protruding}{numeric, modifies aspect of cM/cMLeft marks, see
parameter
\code{protruding} in \code{\link{plotIdiograms}}, optional}
}

An object of class \code{data.frame} with 395 rows and 3 columns.

An object of class \code{data.frame} with 6 rows and 3 columns.

An object of class \code{data.frame} with 2 rows and 3 columns.
}
\source{
\url{http://www.pathology.washington.edu/research/cytopages/idiograms/human/}
}
\usage{
dfMarkColor

humMarkColor

mydfMaColor

dfMarkColor5S25S
}
\description{
style column does not apply to cen. marks, only color.

dfMarkColor: Example General data for marks NOT position

humMarkColor: human bands' characteristics, from Adler (1994)

mydfMaColor: mark characteristics used in vignette of phylogeny

dfMarkColor5S25S: mark characteristics used in circular plot vignette
}
\references{
Adler 1994. Idiogram Album. URL:
\url{http://www.pathology.washington.edu/research/cytopages/idiograms/human/}
}
\seealso{
\code{\link{plotIdiograms}}

\code{\link{markposDFs}}

\code{\link{markdataholo}}
}
\keyword{datasets}
