% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfMarkStyle.R
\docType{data}
\name{dfMarkStyle}
\alias{dfMarkStyle}
\alias{dfMarkColor}
\alias{humMarkColor}
\alias{mydfMaColor}
\title{Mark characteristics}
\format{dfMarkColor a data.frame with columns:
\describe{
\item{markName}{name of mark}
\item{markColor}{use R colors}
\item{style}{character, use square or dots, optional}
\item{protruding}{numeric, modifies aspect of cM/cMLeft marks, see parameter \code{protruding} in \code{\link{plotIdiograms}}, optional}
}}
\source{
\url{http://www.pathology.washington.edu/research/cytopages/idiograms/human/}
}
\usage{
dfMarkColor

humMarkColor

mydfMaColor
}
\description{
style column does not apply to cen. marks, only color.

dfMarkColor: Example General data for marks NOT position

humMarkColor: human bands' characteristics, from Adler (1994)

mydfMaColor: mark characteristics used in vignette of phylogeny
}
\references{
Adler 1994. Idiogram Album. URL: \url{http://www.pathology.washington.edu/research/cytopages/idiograms/human/}
}
\seealso{
\code{\link{plotIdiograms}}

\code{\link{markposDFs}}

\code{\link{markdataholo}}
}
\keyword{datasets}
