% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genBankReadIF.R
\name{genBankReadIF}
\alias{genBankReadIF}
\title{FUNCTION genBankReadIF}
\usage{
genBankReadIF(filename.gb)
}
\arguments{
\item{filename.gb}{name of file to read, downloaded from genBank}
}
\value{
list
}
\description{
genBankReadIF: creates a list of data.frames from a genbank format file
}
\keyword{genBank}
