% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/idiogramFISH-package.R
\docType{package}
\name{idiogramFISH-package}
\alias{idiogramFISH}
\alias{idiogramFISH-package}
\title{idiogramFISH: Shiny App. Idiograms with Marks and Karyotype Indices}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Plot idiograms of karyotypes, plasmids, circular chr. having a set of data.frames for chromosome data and optionally mark data. Two styles of chromosomes can be used: without or with visible chromatids (when not circular). Supports micrometers, cM and Mb or any unit. Two styles of centromeres are available: triangular and rounded; and six styles of marks: square (squareLeft), dots, cM (cMLeft), cenStyle, upArrow, downArrow; its legend (label) can be drawn inline or to the right of karyotypes. Idiograms can also be plotted in concentric circles. It is possible to calculate chromosome indices by Levan et al. (1964) <doi:10.1111/j.1601-5223.1964.tb01953.x>, karyotype indices of Watanabe et al. (1999) <doi:10.1007/PL00013869> and Romero-Zarco (1986) <doi:10.2307/1221906> and classify chromosomes by morphology Guerra (1986) and Levan et al. (1964).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ferroao.gitlab.io/manualidiogramfish/}
  \item \url{https://ferroao.gitlab.io/idiogramFISH}
  \item Report bugs at \url{https://gitlab.com/ferroao/idiogramFISH/-/issues}
}

}
\author{
\strong{Maintainer}: Fernando Roa \email{froao@unal.edu.co}

Other contributors:
\itemize{
  \item Mariana PC Telles [contributor]
}

}
\keyword{internal}
