% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makedfMarkColor.R
\name{makedfMarkColor}
\alias{makedfMarkColor}
\title{Function to automatically make a data.frame of Marks' characterisitcs}
\usage{
makedfMarkColor(
  dfMarkColor,
  markNames,
  colorstoremove = NA,
  defaultStyleMark = "square"
)
}
\arguments{
\item{dfMarkColor}{d.f of marks to use}

\item{markNames}{names of marks}

\item{colorstoremove}{remove this colors from the result}

\item{defaultStyleMark}{character vector with default style \code{"square"} \code{"cM"} \code{"dots"}}
}
\value{
data.frame
}
\description{
This function reads character vector with marks names

It returns a data.frame with color and style for each mark
}
\keyword{internal}
\keyword{mark}
