% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makedfMarkColorMycolors.R
\name{makedfMarkColorMycolors}
\alias{makedfMarkColorMycolors}
\title{Function to automatically make a data.frame of Marks' characterisitcs}
\usage{
makedfMarkColorMycolors(
  markNames,
  mycolors,
  colorstoremove = NULL,
  defaultStyleMark = "square"
)
}
\arguments{
\item{markNames}{names of marks}

\item{mycolors}{character vector of names of colors}

\item{colorstoremove}{character vector of colors to remove from mycolors}

\item{defaultStyleMark}{character vector with default style \code{"square"}
\code{"cM"}
\code{"dots"}}
}
\value{
data.frame
}
\description{
This function reads character vector with marks names, a
character vector of
prohibited colors and a custom list of colors
to be associated with the mark names

It returns a data.frame with color and style for each mark
}
\keyword{internal}
\keyword{mark}
