% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeNumCols.R
\name{makeNumCols}
\alias{makeNumCols}
\title{This is an internal function that eliminates factors}
\usage{
makeNumCols(df)
}
\arguments{
\item{df}{data.frame}
}
\value{
data.frame
}
\description{
It returns a data.frames
}
\keyword{internal}
