% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapXYCenMimic.R
\name{mapXYCenMimic}
\alias{mapXYCenMimic}
\title{mapXYCenMimic
This is an internal function that creates coords for chr. or arms or marks}
\usage{
mapXYCenMimic(start, end, y, x, yfactor, r2, pts, cenFormat = "triangle")
}
\arguments{
\item{start}{1st index}

\item{end}{last index}

\item{y}{y coords}

\item{yfactor}{y x factor}

\item{r2}{r2ius}

\item{pts}{points of squareness}

\item{cenFormat}{boolean}
}
\value{
list
}
\description{
mapXYCenMimic
This is an internal function that creates coords for chr. or arms or marks
}
\keyword{internal}
