% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markdataholo.R
\docType{data}
\name{markdataholo}
\alias{markdataholo}
\alias{bigdfMarkPosHolo}
\alias{dfMarkPosHolo}
\title{Mark data Holo.}
\format{data.frame with cols:
\describe{
\item{OTU}{OTU, species, optional}
\item{chrName}{name of chromosome}
\item{markName}{name of mark}
\item{markPos}{position from bottom or top (see parameter origin in plotIdiogramsHolo)}
\item{markSize}{size of mark in micrometers or Mb}
\item{group}{chr. group, optional}
}}
\usage{
bigdfMarkPosHolo

dfMarkPosHolo
}
\description{
bigdfMarkPosHolo: Example data for mark position of holocentrics with column OTU

dfMarkPosHolo: Example data for mark position of holocentrics
}
\seealso{
\code{\link{plotIdiogramsHolo}}

\code{\link{chrbasicdataHolo}}
}
\keyword{datasets}
