% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perMark.R
\name{perMark}
\alias{perMark}
\title{FUNCTION perMark}
\usage{
perMark(dfMarkPos, listOfdfChromSize)
}
\arguments{
\item{dfMarkPos}{data.frame of marks' position}

\item{listOfdfChromSize}{list data.frames of chr. sizes. Require col. \code{chrSize} for all. Use \code{armRatioCI}}
}
\value{
list
}
\description{
calculates fraction (\%) of chromosome for each mark
}
\examples{
# Beginning with two data.frames with column OTU
dfOfChrSize$OTU<-1
dfOfMarks2$OTU<-1

# add column chrSize if missing
dfOfChrSizeWithChrSize <- armRatioCI(dfOfChrSize)

#data.frame of chr. size to list
listOfdfChr<-idiogramFISH:::dfToListColumn(dfOfChrSizeWithChrSize)

perMark(dfOfMarks2,listOfdfChr)

}
\keyword{chromosome}
\keyword{mark}
\keyword{percentage}
\keyword{span}
