% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roundPlotMark.R
\name{roundPlotMark}
\alias{roundPlotMark}
\title{This is an internal function that eliminates factors}
\usage{
roundPlotMark(roundness, x, y, dfMarkColorInternal, listOfdfMarkPosSq,
  chrWidth, yfactor, n2, lwd.chr)
}
\arguments{
\item{roundness}{roundness of vertices <}

\item{x}{x component of polygon}

\item{y}{y component of polygon}

\item{dfMarkColorInternal}{colors for marks}

\item{listOfdfMarkPosSq}{list of df. of mark pos.}

\item{chrWidth}{width of chr.}

\item{yfactor}{y distortion based on canvas proportion}

\item{n2}{numeric, to define vertices of rounded portions}

\item{lwd.chr}{thick of border line}
}
\value{
plot
}
\description{
It returns a dataframes
}
\keyword{internal}
