% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulerPlot.R
\name{rulerPlot}
\alias{rulerPlot}
\title{rulerPlot
This is an internal function that plots rulers}
\usage{
rulerPlot(
  ycoord,
  listOfdfChromSize,
  listOfdfChromSizeCenType,
  fromZerotoMax,
  rulerNumberSize,
  rulerPosMod,
  rulerPos,
  ruler.tck,
  lwd.chr
)
}
\arguments{
\item{ycoord}{y coords}

\item{listOfdfChromSize}{list of all d.f chr size}

\item{listOfdfChromSizeCenType}{only d.f. of this cen. type}

\item{fromZerotoMax}{intervals}

\item{rulerNumberSize}{font}

\item{rulerPosMod}{modifier of pos.}

\item{rulerPos}{pos. of ruler}

\item{ruler.tck}{tick size and orient.}
}
\value{
axis
}
\description{
It returns rulers (axis)
}
\keyword{internal}
