% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rulerTitle.R
\name{rulerTitle}
\alias{rulerTitle}
\title{rulerTitle
This is an internal function that plots rulers}
\usage{
rulerTitle(
  ycoord,
  listOfdfChromSizeCenType,
  MbUnit,
  specialyTitle,
  yTitle,
  ylabline
)
}
\arguments{
\item{ycoord}{y coords}

\item{listOfdfChromSizeCenType}{only d.f. of this cen. type}

\item{MbUnit}{intervals}

\item{specialyTitle}{modifier of pos.}

\item{yTitle}{pos. of ruler}

\item{ylabline}{font}
}
\value{
plot
}
\description{
It returns rulers (axis)
}
\keyword{internal}
