% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orientation.R
\name{yVertoHor}
\alias{yVertoHor}
\title{FUNCTIONS: yVertoHor, xHortoVer, xHortoVerRoundCen, xHortoVerDots,
mapCircle, mapRadius,
transYList, mapOTUnames, addOTUnames, applyMapCircle, intercalate,
drawPlot, drawPlotMark,
drawPlotMarkLine, drawCen, circLabelMark, circPlotDots, plotChrNames,
addChrNameAttrMark,
addChrNameAttrMarkDots, xMarkMap, xMarkMapLeft, markMapPer, markMapPercM,
markMapPerDots,
radDotsPer, centerMarkMapPer, transyListMark, transyListCen, transRadDots,
transyListMarkDots,
oneDot, OTUlabelsright}
\usage{
yVertoHor(y, monocenNames)
}
\arguments{
\item{y}{list, chr coords.}

\item{monocenNames}{character, vector with names of OTUs}

\item{ylist}{y coords}

\item{x}{only d.f. of this cen. type}

\item{r3}{modifier of pos.}

\item{circleCenterY}{numeric}

\item{cenBorder}{character, cen border color}

\item{cenColor2}{character, cen. color}

\item{cfunction}{character, name of function to apply}

\item{lwd.chr}{numeric, width of line}

\item{chrColor}{character, chr. color}

\item{chrId}{character, type of chr. name}

\item{chrNames}{list, resulting from applyMapCircle}

\item{chrWidth}{character, chr. width}

\item{circleCenter}{numeric, modify center of circle of chr.}

\item{circleMaps}{list, resulting from applyMapCircle}

\item{circleMapsMarksCr}{list, resulting from applyMapCircle}

\item{circleMapsOTUname}{list, resulting from applyMapCircle}

\item{colBorderCr}{character, border color}

\item{colCr}{character color of mark}

\item{dfMarkColorInternal}{data.frame of mark charact.}

\item{diffXRounded}{numeric, difference of chr. width to rounded corner}

\item{firstXchrEach}{list, first element from xlistNewChr}

\item{firstYchrEach}{list, first element from xlistNewChr}

\item{fixCenBorder2}{boolean, change cen border}

\item{indexIdTextSize}{numeric, text size}

\item{labelSpacing}{numeric, spacing of labels to chr. or mark.}

\item{parparlistOfdfMarkPosDataCen}{list, of d.f.s of marks}

\item{listOfdfMarkPos}{list, of d.f.s of marks}

\item{listOfdfMarkPosSq}{list, of d.f.s of marks}

\item{mapCircle}{function name}

\item{markDistType}{character, type of distance to mark}

\item{markLabelSize}{numeric, size of mark label}

\item{n}{numeric, number of vertices}

\item{OTUplacing}{boolean use number instead of OTU name}

\item{OTUlabelSpacing}{numeric, spacing for OTU name}

\item{OTUsrt}{numeric, angle for OTU name}

\item{OTUTextSize}{numeric, font size of OTU}

\item{pattern}{character, regex pattern to remove from mark names}

\item{position}{numeric, indexer for karyotype position}

\item{radius}{numeric, radius}

\item{rad}{list, list of radius for dots marks}

\item{radiusMap}{function, map radius in circle}

\item{radPerCr}{list, radius percentages}

\item{separFactor}{numeric, separation of kar. in circle}

\item{shrinkFactor}{numeric, percentage in decimal of whole chr.}

\item{unlist}{boolean, for unlisting x or y coordinates}

\item{xfactor}{numeric, modify y x aspect}

\item{xlist}{list, x coordinates}

\item{xlistMarkCen}{list, marks}

\item{xlistNew}{list, transformed coordinates}

\item{xlistNewCen}{list, transformed coordinates}

\item{xlistNewChr}{list, transformed coordinates}

\item{xMark}{list, mark coords.}

\item{xMarkCr}{list, mark coords.}

\item{xMarkList}{list, mark coords.}

\item{xRounded}{list, transformed coords.}

\item{ylistNew}{list, chr coords.}

\item{ylistNewCen}{list, chr coords.}

\item{ylistNewChr}{list, chr coords.}

\item{ylistTrans}{list, trans .chr coords.}

\item{ylistTransChr}{list, trans .chr coords.}

\item{yMark}{list, mark coords.}

\item{yMarkCen}{list, mark coords.}

\item{yMarkPer}{list, mark coords in percentage.}

\item{OTUlegendHeight}{numeric}

\item{labelOutwards}{srt}
}
\value{
plot
}
\description{
FUNCTIONS: yVertoHor, xHortoVer, xHortoVerRoundCen, xHortoVerDots,
mapCircle, mapRadius,
transYList, mapOTUnames, addOTUnames, applyMapCircle, intercalate,
drawPlot, drawPlotMark,
drawPlotMarkLine, drawCen, circLabelMark, circPlotDots, plotChrNames,
addChrNameAttrMark,
addChrNameAttrMarkDots, xMarkMap, xMarkMapLeft, markMapPer, markMapPercM,
markMapPerDots,
radDotsPer, centerMarkMapPer, transyListMark, transyListCen, transRadDots,
transyListMarkDots,
oneDot, OTUlabelsright
}
\keyword{internal}
